Redesigned Move Interface
Now Move Interface is even more encapsulated and object oriented!
All what Player has to do is to call ReceiveStep() method on
GameElement he wants to step on. The method will return green or
red light to go, and GameElement itself will decide what to do or not
to do with the player itself.